MNet: fixed wrong column name in SQL cleanup code

+++ auth/mnet/auth.php
@@ -549,9 +549,11 @@ class auth_plugin_mnet extends auth_plugin_base {
         }
 
         // Clean up courses that the user is no longer enrolled in.
-        $local_courseid_string = implode(', ', $local_courseid_array);
-        $whereclause = " userid = ? AND hostid = ? AND courseid NOT IN ($local_courseid_string)";
-        $DB->delete_records_select('mnetservice_enrol_enrolments', $whereclause, array($userid, $remoteclient->id));
+        if (!empty($local_courseid_array)) {
+            $local_courseid_string = implode(', ', $local_courseid_array);
+            $whereclause = " userid = ? AND hostid = ? AND remotecourseid NOT IN ($local_courseid_string)";
+            $DB->delete_records_select('mnetservice_enrol_enrolments', $whereclause, array($userid, $remoteclient->id));
+        }
     }
 
     function prevent_local_passwords() {
